Conditional rules cleanup
