+++ src/RuleEngine.Tests/Rules/ConditionalRulesTests.cs
@@ -1,8 +1,9 @@
-﻿using System;
-using System.Linq.Expressions;
-using FluentAssertions;
+﻿using FluentAssertions;
 using RuleEngine.Rules;
+using RuleEngine.Rules.Conditional;
 using SampleModel;
+using System;
+using System.Linq.Expressions;
 using Xunit;
 using Xunit.Abstractions;
