+++ src/RuleFactory.Tests/JsonRules/ConditionalRulesJsonTests.cs
@@ -1,8 +1,9 @@
-﻿using System;
-using FluentAssertions;
+﻿using FluentAssertions;
 using Newtonsoft.Json;
 using RuleEngine.Rules;
+using RuleEngine.Rules.Conditional;
 using SampleModel;
+using System;
 using Xunit;
 using Xunit.Abstractions;
