+++ src/RuleFactory/RuleFactory.cs
@@ -1,7 +1,7 @@
-﻿using System;
-using RuleEngine.Rules;
+﻿using RuleEngine.Rules;
+using RuleEngine.Rules.Conditional;
 using RuleEngine.Utils;
-//using RuleFactory.Factory;
+using System;
 
 namespace RuleFactory
 {
