+++ src/RuleFactory/RulesFactory/ConditionalRulesFactory.cs
@@ -1,4 +1,5 @@
 ﻿using RuleEngine.Rules;
+using RuleEngine.Rules.Conditional;
 
 namespace RuleFactory.RulesFactory
 {
